Moving some code to the utils file
